Move inchiparsingTest contents into translatorTest
